Create SessionLinearLayoutManager for save and restore position
